Refactoring of hand ray
- Created IHandRay interface.
- Extracted StabilizedRay to utilities class.
- Cleaned up HandRay code.
