+++ Assets/MixedRealityToolkit/Providers/Hands/HandRay.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 458f4d0fc0e513b4993b24828b73742e
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
