+++ Assets/MixedRealityToolkit/Providers/Hands/HandRay/HandRay.cs
@@ -6,8 +6,9 @@ using UnityEngine;
 
 namespace Microsoft.MixedReality.Toolkit.Input
 {
-    public class HandRay
+    public class HandRay : IHandRay
     {
+        /// <inheritdoc />
         public Ray Ray
         {
             get
@@ -18,6 +19,7 @@ namespace Microsoft.MixedReality.Toolkit.Input
             }
         }
 
+        /// <inheritdoc />
         public bool ShouldShowRay
         {
             get
@@ -49,7 +51,7 @@ namespace Microsoft.MixedReality.Toolkit.Input
 
         private Ray ray = new Ray();
 
-        // Constants from Shell Implementation of hand ray
+        // Constants from Shell Implementation of hand ray.
         private const float DynamicPivotBaseY = -0.1f, DynamicPivotMultiplierY = 0.65f, DynamicPivotMinY = -0.6f, DynamicPivotMaxY = -0.2f;
         private const float DynamicPivotBaseX = 0.03f, DynamicPivotMultiplierX = 0.65f, DynamicPivotMinX = 0.08f, DynamicPivotMaxX = 0.15f;
         private const float HeadToPivotOffsetZ = 0.08f;
@@ -57,15 +59,15 @@ namespace Microsoft.MixedReality.Toolkit.Input
         private readonly float CursorBeamUpTolerance = 0.8f;
 
 
-        // Smoothing factor for ray stabilization
+        // Smoothing factor for ray stabilization.
         private const float StabilizedRayHalfLife = 0.01f;
 
-        private StabilizedRay stabilizedRay = new StabilizedRay(StabilizedRayHalfLife);
+        private readonly StabilizedRay stabilizedRay = new StabilizedRay(StabilizedRayHalfLife);
         private Vector3 palmNormal;
         private Vector3 headForward;
 
         #region Public Methods
-
+        /// <inheritdoc />
         public void Update(Vector3 handPosition, Vector3 palmNormal, Transform headTransform, Handedness sourceHandedness)
         {
             Vector3 rayPivotPoint = ComputeRayPivotPosition(handPosition, headTransform, sourceHandedness);
@@ -76,7 +78,6 @@ namespace Microsoft.MixedReality.Toolkit.Input
 
             stabilizedRay.AddSample(new Ray(measuredRayPosition, measuredDirection));
         }
-
         #endregion
 
 
@@ -108,82 +109,6 @@ namespace Microsoft.MixedReality.Toolkit.Input
 
             Quaternion headRotationFlat = Quaternion.Euler(0, headTransform.rotation.eulerAngles.y, 0);
             return headTransform.position + headRotationFlat * relativePivot;
-        }
-    }
-
-    /// <summary>
-    /// A ray whose position and direction are stabilized in a way similar to how gaze stabilization
-    /// works in HoloLens.
-    /// 
-    /// The ray uses Anatolie Gavrulic's "DynamicExpDecay" filter to stabilize the ray
-    /// this filter adjusts its smoothing factor based on the velocity of the filtered object
-    /// 
-    /// The formula is
-    ///  Y_smooted += ∆𝑌_𝑟
-    ///  where
-    /// 〖∆𝑌_𝑟=∆𝑌∗〖0.5〗^(∆𝑌/〖Halflife〗)
-    /// 
-    /// In code, LERP(signal, oldValue, POW(0.5, ABS(signal – oldValue) / hl)
-    /// </summary>
-    internal class StabilizedRay
-    {
-        public float HalfLifePosition { get; } = 0.1f;
-
-        public float HalfLifeDirection { get; } = 0.1f;
-
-        public Vector3 StabilizedPosition { get; private set; }
-
-        public Vector3 StabilizedDirection { get; private set; }
-
-        public Quaternion StabilizedRotation => Mathf.Abs(StabilizedDirection.magnitude) < 1e-5f ?
-            Quaternion.identity : Quaternion.LookRotation(StabilizedDirection);
-
-        private bool isInitialized = false;
-
-        /// <summary>
-        /// HalfLife closer to zero means lerp closer to one
-        /// </summary>
-        public StabilizedRay(float halfLife)
-        {
-            HalfLifePosition = halfLife;
-            HalfLifeDirection = halfLife;
-        }
-
-        public StabilizedRay(float halfLifePos, float halfLifeDir)
-        {
-            HalfLifePosition = halfLifePos;
-            HalfLifeDirection = halfLifeDir;
-        }
-
-        public void AddSample(Ray ray)
-        {
-            if (!isInitialized)
-            {
-                StabilizedPosition = ray.origin;
-                StabilizedDirection = ray.direction.normalized;
-                isInitialized = true;
-            }
-            else
-            {
-                StabilizedPosition = DynamicExpDecay(StabilizedPosition, ray.origin, HalfLifePosition);
-                StabilizedDirection = DynamicExpDecay(StabilizedDirection, ray.direction.normalized, HalfLifeDirection);
-            }
-        }
-
-
-        public static float DynamicExpCoefficient(float hLife, float delta)
-        {
-            if (hLife == 0)
-            {
-                return 1;
-            }
-
-            return 1.0f - Mathf.Pow(0.5f, delta / hLife);
-        }
-
-        public static Vector3 DynamicExpDecay(Vector3 from, Vector3 to, float hLife)
-        {
-            return Vector3.Lerp(from, to, DynamicExpCoefficient(hLife, Vector3.Distance(to, from)));
         }
     }
 }
