+++ Assets/MixedRealityToolkit/Providers/Hands/HandRay/IHandRay.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 78d846f555e0d17488fab3d66710b3a7
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
