+++ Assets/MixedRealityToolkit/Utilities/StabilizedRay.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e966538be6393884d857b4045424be13
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
